requirements.txt
transformers
torch
streamlit
openai-whisper
